Clean up SQLite lock error handling
